update layout and finish up

+++ app/src/main/java/com/example/simpleMovieListing/model/Movie.kt
@@ -3,7 +3,7 @@ package com.example.simpleMovieListing.model
 data class Movie(
     var uuid: String,
     var name: String,
-    var openDate: Int,
+    var openDate: Long,
     var poster: String,
     var rating: Float,
     var likeCount: Int,
